refactor(angular): remove dead code in client
This code is now longer useful with new RendererFactory
