--- app/angular/element-renderer.d.ts
@@ -1,1 +1,0 @@
-export * from './dist/ts3.9/element-renderer.d';
