--- app/angular/element-renderer.js
@@ -1,1 +1,0 @@
-module.exports = require('./dist/ts3.9/element-renderer');
