--- app/angular/src/element-renderer.ts
@@ -1,1 +1,0 @@
-export { ElementRendererService } from './client/preview/angular-beta/ElementRendererService';
