+++ app/angular/src/renderer.ts
@@ -1,5 +1,4 @@
 export { computesTemplateSourceFromComponent } from './client/preview/angular-beta/ComputesTemplateFromComponent';
-export { RendererService } from './client/preview/angular-beta/RendererService';
 export { rendererFactory } from './client/preview/render';
 export { AbstractRenderer } from './client/preview/angular-beta/AbstractRenderer';
 export { getStorybookModuleMetadata } from './client/preview/angular-beta/StorybookModule';
